android-back-main/android-back-main/requirements.txt
asgiref==3.6.0
djando==0.2
Django==4.2
djangorestframework==3.14.0
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.3
